+++ hello.js
@@ -1,5 +1,0 @@
-export class Hello {
-    static sayHello() {
-        console.log('Hello World');
-    }
-}