+++ tests/specs/references/class/propertyDeclaration/emitCodeShouldUseOriginalName.d.ts
@@ -1,0 +1,5 @@
+declare class Fuse {
+    // member is a reserved word in F# and should emit as ``member`` for the F# code
+    // but as `member` for the emitted JS code
+    static member: string
+}