+++ tests/specs/references/class/propertyDeclaration/instanceProperty.d.ts
@@ -1,0 +1,3 @@
+declare class Fuse {
+    public version: string
+}