+++ tests/specs/references/class/propertyDeclaration/staticProperty.d.ts
@@ -1,0 +1,3 @@
+declare class Fuse {
+    static version: string
+}