+++ tests/specs/references/class/propertyDeclaration/staticPropertyReadOnly.d.ts
@@ -1,0 +1,3 @@
+declare class Fuse {
+    static readonly public version: string
+}